requirements.txt
# Core Framework
flet[all]>=0.28.3  # Latest stable version (May 20, 2025) - supports Python 3.9-3.13

# Database
pandas>=2.1.0

# Telegram API
telethon>=1.34.0

# Firebase Authentication (using REST API - no Admin SDK needed in desktop app)
firebase-admin>=6.2.0  # Only needed for deployment scripts and admin interface, not desktop app
PyJWT>=2.8.0  # For decoding Firebase ID tokens
google-cloud-firestore>=2.11.0  # For gRPC-based real-time watch streams

# Export Libraries
xlsxwriter>=3.1.9
openpyxl>=3.1.2  # For Excel export (used by admin interface)
reportlab>=4.0.7
Pillow>=10.1.0
qrcode[pil]>=7.4.2

# Utilities
python-dotenv>=1.0.0
cryptography>=41.0.5
requests>=2.31.0
beautifulsoup4>=4.14.0
httpx[http2]>=0.25.0  # For Firestore real-time watch streams

# Platform-specific Authentication
pywin32>=306; sys_platform == "win32"
python-pam>=2.0.0; sys_platform == "linux"

# Build
pyinstaller>=6.1.0
pyarmor>=8.5.0  # Code obfuscation and protection

# Testing
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0

